requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==4.0.3
djangorestframework==3.13.1
flake8==4.0.1
mccabe==0.6.1
mypy==0.941
mypy-extensions==0.4.3
Pillow==9.0.1
pkg_resources==0.0.0
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
